+++ src/main/java/com/zerobase/munbanggu/user/service/AuthService.java
@@ -1,6 +1,6 @@
 package com.zerobase.munbanggu.user.service;
 
-import com.zerobase.munbanggu.config.auth.TokenProvider;
+import com.zerobase.munbanggu.auth.TokenProvider;
 import com.zerobase.munbanggu.util.RedisUtil;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
